requirements.txt
# Core framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# LangChain & LangGraph ecosystem
langchain>=0.3.0
langchain-openai>=0.1.0
langgraph>=0.2.0
langchain-community>=0.3.0

# Redis async client
redis>=5.0.0

# Data & parsing
pydantic>=2.7.0
pandas>=2.2.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0
python-dotenv>=1.0.0

# Optional (if using asyncio run tools etc.)
# ast is in stdlib; subprocess/tempfile/base64 also stdlib.

# NOTE: Azure OpenAI SDK not required separately; langchain-openai handles API bindings.
